create test_formats.py and move tests
